requirements.txt
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord==1.0.1
discord.py==1.7.3
dnspython==2.1.0
idna==2.10
multidict==5.1.0
nextcord==2.0.0a2
pymongo==3.11.4
python-dotenv==0.19.2
pytz==2021.1
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
